feat: widen DeviceState.battery from int to float
- zigbee2mqtt reports fractional battery percentages (e.g. 91.5%);
  rounding to int truncated those at the adapter boundary. Domain type,
  adapter parsing, alarm low-battery message, activity format, and
  history recorder all carry float now. Frontend cards and sensor
  display round only at the rendering edge with `Math.round`.

+++ internal/adapter/zigbee/dto.go
@@ -35,7 +35,7 @@ type z2mDeviceState struct {
 
 	Temperature *float64 `json:"temperature"`
 	Humidity    *float64 `json:"humidity"`
-	Battery     *int     `json:"battery"`
+	Battery     *float64 `json:"battery"`
 	Pressure    *float64 `json:"pressure"`
 	Illuminance *float64 `json:"illuminance"`
